Navigation Utilities
- update snapshot type
- Migrate navigation utilities to own folder
