+++ libtelemetry/gradle.properties
@@ -1,4 +1,4 @@
-VERSION_NAME=4.0.0-SNAPSHOT
+VERSION_NAME=3.1.0-SNAPSHOT
 POM_ARTIFACT_ID=mapbox-android-telemetry
 POM_NAME=Mapbox Android Telemetry Library
 POM_DESCRIPTION=Mapbox Android Telemetry Library
